+++ backend/src/main/java/com/yukgaejang/voss/domain/freeboard/service/PostServiceImpl.java
@@ -1,8 +1,12 @@
 package com.yukgaejang.voss.domain.freeboard.service;
 
+import com.yukgaejang.voss.domain.freeboard.exception.NoPostCommentException;
 import com.yukgaejang.voss.domain.freeboard.exception.NoPostException;
+import com.yukgaejang.voss.domain.freeboard.repository.PostCommentRepository;
+import com.yukgaejang.voss.domain.freeboard.repository.PostLikeRepository;
 import com.yukgaejang.voss.domain.freeboard.repository.PostRepository;
 import com.yukgaejang.voss.domain.freeboard.repository.entity.Post;
+import com.yukgaejang.voss.domain.freeboard.repository.entity.PostComment;
 import com.yukgaejang.voss.domain.freeboard.service.PostService;
 import com.yukgaejang.voss.domain.freeboard.service.dto.request.CreatePostRequest;
 import com.yukgaejang.voss.domain.freeboard.service.dto.request.UpdatePostRequest;
@@ -12,10 +16,12 @@ import com.yukgaejang.voss.domain.member.repository.MemberRepository;
 import com.yukgaejang.voss.domain.member.repository.entity.Member;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
+import java.util.List;
 import java.util.Optional;
 
 @Service
@@ -23,6 +29,8 @@ import java.util.Optional;
 public class PostServiceImpl implements PostService {
 
     private final PostRepository postRepository;
+    private final PostCommentRepository postCommentRepository;
+    private final PostLikeRepository postLikeRepository;
     private final MemberRepository memberRepository;
 
     @Override
@@ -35,7 +43,10 @@ public class PostServiceImpl implements PostService {
 
     @Override
     public UpdatePostResponse updatePost(Long id, UpdatePostRequest updatePostRequest) {
-        Post post = postRepository.findById(id).orElseThrow(() -> new NoPostException("존재하지 않는 글입니다."));
+        Post post = postRepository.findByIdAndIsDeletedFalse(id);
+        if(post == null) {
+            throw new NoPostException("존재하지 않는 글입니다.");
+        }
         post.updatePost(updatePostRequest.getTitle(), updatePostRequest.getContent());
         postRepository.save(post);
         return new UpdatePostResponse(true);
@@ -43,22 +54,35 @@ public class PostServiceImpl implements PostService {
 
     @Override
     public PostDetailResponse getPostDetail(Long id) {
-        Post post = postRepository.findById(id).orElseThrow(() -> new NoPostException("존재하지 않는 글입니다."));
+        Post post = postRepository.findByIdAndIsDeletedFalse(id);
+        if(post == null) {
+            throw new NoPostException("존재하지 않는 글입니다.");
+        }
         post.updateHit();
-        return new PostDetailResponse(postRepository.save(post));
+        Page<CommentDetailResponse> comments = new PageImpl<>(postCommentRepository.findAllByPostIdAndIsDeletedFalse(id));
+        Long likes = postLikeRepository.countByPostId(id);
+        return new PostDetailResponse(postRepository.save(post), comments, likes);
     }
 
     @Override
     public Page<PostListResponse> getPostList(Pageable pageable) {
-        Page<Post> allPosts = postRepository.findAllByIsDeletedFalse(pageable);
-        return allPosts.map(PostListResponse::new);
+        return postRepository.findAllByIsDeletedFalse(pageable);
     }
 
     @Override
     public DeletePostResponse deletePost(Long id) {
-        Post post = postRepository.findById(id).orElseThrow(() -> new NoPostException("존재하지 않는 글입니다."));
+        Post post = postRepository.findByIdAndIsDeletedFalse(id);
+        if(post == null) {
+            throw new NoPostException("존재하지 않는 글입니다.");
+        }
         post.delete();
         postRepository.save(post);
+        List<CommentDetailResponse> comments = postCommentRepository.findAllByPostIdAndIsDeletedFalse(id);
+        for(CommentDetailResponse comment : comments) {
+            PostComment postComment = postCommentRepository.findById(comment.getId()).orElseThrow(() -> new NoPostCommentException("존재하지 않는 댓글입니다."));
+            postComment.delete();
+            postCommentRepository.save(postComment);
+        }
         return new DeletePostResponse(true);
     }
